Move the base device wrapper into a util class
